# spec_snapshot/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadCleanupCandidate": {
        "properties": {
          "age_seconds": {
            "title": "Age Seconds",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "name",
          "age_seconds"
        ],
        "title": "UploadCleanupCandidate",
        "type": "object"
      },
      "UploadCleanupDeleteRequest": {
        "properties": {
          "dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dir"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "older_than": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Older Than"
          },
          "pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern"
          }
        },
        "title": "UploadCleanupDeleteRequest",
        "type": "object"
      },
      "UploadCleanupDeleteResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "deleted": {
            "items": {
              "type": "string"
            },
            "title": "Deleted",
            "type": "array"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/UploadCleanupError"
            },
            "title": "Errors",
            "type": "array"
          }
        },
        "required": [
          "deleted",
          "errors",
          "count"
        ],
        "title": "UploadCleanupDeleteResponse",
        "type": "object"
      },
      "UploadCleanupError": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "error"
        ],
        "title": "UploadCleanupError",
        "type": "object"
      },
      "UploadCleanupListResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/UploadCleanupCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "candidates",
          "count"
        ],
        "title": "UploadCleanupListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Minutes Service (prototype)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/uploads/cleanup": {
      "get": {
        "operationId": "list_uploads_for_cleanup_admin_uploads_cleanup_get",
        "parameters": [
          {
            "in": "query",
            "name": "dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dir"
            }
          },
          {
            "in": "query",
            "name": "pattern",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pattern",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "older_than",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Older Than",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCleanupListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "List Uploads For Cleanup",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "delete_uploads_admin_uploads_cleanup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCleanupDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCleanupDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Delete Uploads",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/uploads/cleanup": {
      "get": {
        "operationId": "list_uploads_for_cleanup_api_admin_uploads_cleanup_get",
        "parameters": [
          {
            "in": "query",
            "name": "dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dir"
            }
          },
          {
            "in": "query",
            "name": "pattern",
            "required": false,
            "schema": {
              "default": "",
              "title": "Pattern",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "older_than",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Older Than",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCleanupListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "List Uploads For Cleanup",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "delete_uploads_api_admin_uploads_cleanup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCleanupDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCleanupDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Delete Uploads",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}